Move to micronaut-function and rename config path accordingly
